integrate `map_async`
