+++ tqdm/utils.py
@@ -1,6 +1,7 @@
 """
 General helpers required for `tqdm.std`.
 """
+from contextlib import contextmanager
 from functools import wraps
 from warnings import warn
 import os
@@ -268,6 +269,20 @@ class CallbackIOWrapper(ObjectWrapper):
             self.wrapper_setattr('read', read)
         else:
             raise KeyError("Can only wrap read/write methods")
+
+
+@contextmanager
+def ensure_lock(tqdm_class, lock_name=""):
+    """get (create if necessary) and then restore `tqdm_class`'s lock"""
+    old_lock = getattr(tqdm_class, '_lock', None)  # don't create a new lock
+    lock = old_lock or tqdm_class.get_lock()  # maybe create a new lock
+    lock = getattr(lock, lock_name, lock)  # maybe subtype
+    tqdm_class.set_lock(lock)
+    yield lock
+    if old_lock is None:
+        del tqdm_class._lock
+    else:
+        tqdm_class.set_lock(old_lock)
 
 
 def _is_utf(encoding):
